Move and adjust files
