+++ libqfieldsync/project_checker.py
@@ -9,10 +9,11 @@ from qfieldsync.libqfieldsync.layer import (
     SyncAction,
     UnsupportedPrimaryKeyError,
 )
-from qfieldsync.libqfieldsync.project import ProjectConfiguration, ProjectProperties
-from qfieldsync.libqfieldsync.utils.file_utils import isascii
 from qgis.core import Qgis, QgsMapLayer, QgsProject, QgsSettings
 from qgis.PyQt.QtCore import QObject
+
+from libqfieldsync.project import ProjectConfiguration, ProjectProperties
+from libqfieldsync.utils.file_utils import isascii
 
 from .offline_converter import ExportType
