+++ libqfieldsync/utils/bad_layer_handler.py
