+++ libqfieldsync/utils/exceptions.py
