+++ libqfieldsync/utils/file_utils.py
