+++ libqfieldsync/utils/logger.py
