+++ libqfieldsync/utils/qgis.py
