+++ libqfieldsync/utils/xml.py
